refactor: rename `@api/core` to `api.core` (#731)
* refactor: rename `@api/core` to `api.core`

* chore: drop unnecessary npx usage

* chore: drop redundant packages property from lerna.json

* revert: apparently knip doesn't like this?

* test: snapshot

* test: fix this? this shouldn't be here wtf
