+++ lerna.json
@@ -1,4 +1,4 @@
 {
-  "packages": ["packages/*"],
+  "$schema": "node_modules/lerna/schemas/lerna-schema.json",
   "version": "6.1.1"
 }
