+++ package.json
@@ -3,18 +3,18 @@
   "private": true,
   "scripts": {
     "alex": "alex .",
-    "build": "npx lerna run build --stream",
+    "build": "lerna run build --stream",
     "build-docs": "node bin/build-markdown.js",
-    "clean": "npx lerna clean",
+    "clean": "lerna clean",
     "lint": "knip && npm run lint:types && npm run lint:js && npm run prettier",
     "lint:js": "eslint . --ext .js,.ts",
     "lint:types": "npm run lint:types --if-present --workspaces",
     "prepare": "husky install",
     "prettier": "prettier --check .",
     "prettier:write": "prettier --check --write .",
-    "publish": "npx lerna publish",
+    "publish": "lerna publish",
     "pretest": "npm run build",
-    "test": "npx lerna run test --stream",
+    "test": "lerna run test --stream",
     "test:smoke": "npm run test:smoke --workspace=packages/api",
     "version": "npx conventional-changelog-cli --pkg lerna.json -i CHANGELOG.md -s && git add CHANGELOG.md"
   },
