+++ packages/api/test/__fixtures__/sdk/operationid-quirks/index.ts
@@ -1,6 +1,6 @@
-import type { ConfigOptions, FetchResponse } from '@api/core'
+import type { ConfigOptions, FetchResponse } from 'api.core';
 import Oas from 'oas';
-import APICore from '@api/core';
+import APICore from 'api.core';
 import definition from '@api/test-utils/definitions/operationid-quirks.json';
 
 class SDK {
