+++ packages/api/test/__fixtures__/sdk/simple-js-cjs/index.js
@@ -3,12 +3,12 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
     return (mod && mod.__esModule) ? mod : { "default": mod };
 };
 var oas_1 = __importDefault(require("oas"));
-var core_1 = __importDefault(require("@api/core"));
+var api_core_1 = __importDefault(require("api.core"));
 var simple_json_1 = __importDefault(require("@api/test-utils/definitions/simple.json"));
 var SDK = /** @class */ (function () {
     function SDK() {
         this.spec = oas_1.default.init(simple_json_1.default);
-        this.core = new core_1.default(this.spec, 'simple-js-cjs/1.0.0 (api/<<package version>>)');
+        this.core = new api_core_1.default(this.spec, 'simple-js-cjs/1.0.0 (api/<<package version>>)');
     }
     /**
      * Optionally configure various options that the SDK allows.
