+++ packages/core/README.md
@@ -1,9 +1,9 @@
-# @api/core
+# api.core
 
 <p align="center">
-  <a href="https://npm.im/@api/core"><img src="https://img.shields.io/npm/v/@api/core?style=for-the-badge" alt="NPM Version"></a>
-  <a href="https://npm.im/@api/core"><img src="https://img.shields.io/node/v/@api/core?style=for-the-badge" alt="Node Version"></a>
-  <a href="https://npm.im/@api/core"><img src="https://img.shields.io/npm/l/@api/core?style=for-the-badge" alt="MIT License"></a>
+  <a href="https://npm.im/api.core"><img src="https://img.shields.io/npm/v/api.core?style=for-the-badge" alt="NPM Version"></a>
+  <a href="https://npm.im/api.core"><img src="https://img.shields.io/node/v/api.core?style=for-the-badge" alt="Node Version"></a>
+  <a href="https://npm.im/api.core"><img src="https://img.shields.io/npm/l/api.core?style=for-the-badge" alt="MIT License"></a>
   <a href="https://github.com/readmeio/api"><img src="https://img.shields.io/github/actions/workflow/status/readmeio/api/ci.yml?branch=main&style=for-the-badge" alt="Build status"></a>
 </p>
